L3/Building AI Agents with LangChain/requirements.txt
langchain>=1.2.0
langchain-classic>=1.0.0
langchain-openai>=1.0.0
langchain-google-genai>=2.0.0
langchain-ollama>=0.3.0
langchain-community>=0.4.0
langgraph>=1.0.0
ddgs>=9.0.0
requests>=2.31.0
python-dotenv>=1.0.0
